+++ CHANGES.md
@@ -5,7 +5,10 @@ Most lines should begin with one of these words:
 *Add*, *Fix*, *Update*, *Change*, *Deprecate*, *Remove*.
 
 ## [Unreleased](https://github.com/sharpjs/Sharp.Diagnostics.Logging/compare/v0.0.0...HEAD)
-(none)
+- Update build targets to .NET Framework 4.6.1 (was 4.5) and .NET Standard 2.0
+- Update test targets to .NET Framework 4.8 and .NET Core 3.1 (was 3.0)
+- Update System.Diagnostics.DiagnosticSource to 6.0.0
+- Modernize project structure
 
 ## [0.1.0](https://github.com/sharpjs/Sharp.Diagnostics.Logging/compare/v0.0.0...v0.1.0)
 - Add dependency: `System.Diagnostics.DiagnosticSource`